requirements.txt
pandas
scikit-learn
openpyxl

deeppavlov
torch>=2.1.2
transformers>=4.39
scipy
numpy 
fasttext-wheel

langchain_openai
langchain_core
langchain
python-dotenv